Move check functions into a new module utils.py
