+++ tests/test_core.py
@@ -353,18 +353,3 @@ class TestXTCPOPC:
         for (key), value in self.ref_OP.items():
             assert key in self.dic_OP.keys()
             assert_almost_equal(value, self.dic_OP[key])
-
-
-    def test_check_def_file(self):
-        """Test for is_allHs_present()."""
-        assert core.is_allHs_present(self.defop, self.dic_lipid, self.dic_Cname2Hnames)
-
-        test_dic = self.dic_Cname2Hnames.copy()
-        # Remove a random carbon
-        del test_dic["C26"]
-        assert not core.is_allHs_present(self.defop, self.dic_lipid, test_dic)
-
-        test_dic2 = self.dic_Cname2Hnames.copy()
-        # Change the number of hydrogens attached to a carbon
-        test_dic2["CA2"] = ('HA21', 'HA22') # should be = ('HA21', 'HA22', 'HA23')
-        assert not core.is_allHs_present(self.defop, self.dic_lipid, test_dic2)
