+++ tests/test_data/Berger_POPC/op_wrong2.def
@@ -1,0 +1,12 @@
+gamma1_1 POPC C1  H11
+gamma1_2 POPC C1  H12
+gamma1_3 POPC C1  H13
+gamma2_1 POPC C2  H21
+gamma2_3 POPC C2  H23
+gamma3_1 POPC C3  H31
+gamma3_2 POPC C3  H32
+gamma3_3 POPC C3  H33
+beta1 POPC C5  H51
+beta2 POPC C5  H52
+alpha1 POPC C6  H61
+alpha2 POPC C6  H62
